+++ Core/Models/ViewEntities/ViewEOSOutstanding.cs
@@ -1,5 +1,8 @@
 ﻿namespace Core.Models.BusinessEntities;
 
+/// <summary>
+/// This view represents the EOS Outstanding events from the VIEW_AllEvents_CURRENT view where the ModifyFlag is null and the event is not Full Released.
+/// </summary>
 public partial record ViewEOSOutstanding
 {
     public int FacilNo { get; set; }
@@ -28,5 +31,5 @@ public partial record ViewEOSOutstanding
 
     public string? OperatorType { get; set; }
 
-    public decimal? ScanDocsNo { get; set; }
+    public int? ScanDocsNo { get; set; }
 }
